Make install process devided

+++ setup.sh
@@ -13,6 +13,14 @@ confirm () {
   done
 }
 
+confirm_installed () {
+  type $1 >/dev/null 2>&1
+  if [ $? -ne 0 ]; then
+    echo "$1 not installed"
+    exit 1
+  fi
+}
+
 symlink () {
   mkdir -p $(dirname "$2")
   rm -f "$2"
@@ -40,24 +48,25 @@ else
   symlink "$cwd/.hyper.linux.js" "$HOME/.hyper.js"
 fi
 
-if confirm 'Install binaries?'; then
+mkdir -p "$HOME/bin"
+
+if confirm 'Install rbenv?'; then
+  git clone https://github.com/sstephenson/rbenv.git ~/.rbenv
+  git clone https://github.com/sstephenson/ruby-build.git ~/.rbenv/plugins/ruby-build
+fi
+
+if confirm 'Install go binaries?'; then
+  confirm_installed curl
+
   source "$HOME/.bashrc"
-  mkdir -p "$HOME/bin"
-
-  packages=(unzip curl vim)
-  for pkg in ${packages[@]}; do
-    type $pkg >/dev/null 2>&1
-    if [ $? -ne 0 ]; then
-      echo "$pkg not installed"
-      exit 1
-    fi
-  done
-
   type go   >/dev/null 2>&1 || (curl -L https://dl.google.com/go/go1.12.4.linux-amd64.tar.gz | sudo tar xz -C /usr/local)
   type dep  >/dev/null 2>&1 || (curl -L https://raw.githubusercontent.com/golang/dep/master/install.sh | sh)
-  type peco >/dev/null 2>&1 || (curl -L https://github.com/peco/peco/releases/download/v0.5.3/peco_linux_amd64.tar.gz | tar xz && mv -f peco_linux_amd64/peco "$HOME/bin" && rm -rf peco_linux_amd64)
-  type memo >/dev/null 2>&1 || (curl -LO https://github.com/mattn/memo/releases/download/v0.0.4/memo_linux_amd64.zip && unzip memo_linux_amd64.zip -d "$HOME/bin" && rm -f memo_linux_amd64.zip)
-  type ghq  >/dev/null 2>&1 || (curl -LO https://github.com/motemen/ghq/releases/download/v0.8.0/ghq_linux_amd64.zip && mkdir -p _ghq && unzip ghq_linux_amd64.zip -d _ghq && mv -f _ghq/ghq "$HOME/bin" && rm -rf _ghq ghq_linux_amd64.zip)
+fi
+
+if confirm 'Install vim plugins?'; then
+  confirm_installed vim
+  confirm_installed curl
+
   type volt >/dev/null 2>&1 || (curl -L https://github.com/vim-volt/volt/releases/download/v0.3.2/volt-v0.3.2-linux-amd64 -o "$HOME/bin/volt" && chmod a+x "$HOME/bin/volt")
 
   $HOME/bin/volt get -u tomasr/molokai
@@ -69,6 +78,15 @@ if confirm 'Install binaries?'; then
   $HOME/bin/volt get -u tpope/vim-markdown
   $HOME/bin/volt get -u itchyny/lightline.vim
   $HOME/bin/volt get -u justinmk/vim-dirvish
+fi
+
+if confirm 'Install other files?'; then
+  confirm_installed unzip
+  confirm_installed curl
+
+  type peco >/dev/null 2>&1 || (curl -L https://github.com/peco/peco/releases/download/v0.5.3/peco_linux_amd64.tar.gz | tar xz && mv -f peco_linux_amd64/peco "$HOME/bin" && rm -rf peco_linux_amd64)
+  type memo >/dev/null 2>&1 || (curl -LO https://github.com/mattn/memo/releases/download/v0.0.4/memo_linux_amd64.zip && unzip memo_linux_amd64.zip -d "$HOME/bin" && rm -f memo_linux_amd64.zip)
+  type ghq  >/dev/null 2>&1 || (curl -LO https://github.com/motemen/ghq/releases/download/v0.8.0/ghq_linux_amd64.zip && mkdir -p _ghq && unzip ghq_linux_amd64.zip -d _ghq && mv -f _ghq/ghq "$HOME/bin" && rm -rf _ghq ghq_linux_amd64.zip)
 
   curl -L https://github.com/git/git/raw/master/contrib/completion/git-prompt.sh -o "$HOME/.git-prompt.sh"
 fi
